LayoutTests/sputnik/Conformance/12_Statement/12.7_The_continue_Statement/S12.7_A8_T1.html

<html>
<head>
<meta charset='utf-8'>
<style>
.pass {
    font-weight: bold;
    color: green;
}
.fail {
    font-weight: bold;
    color: red;
}
</style>

<script>
if (window.testRunner)
    testRunner.dumpAsText();

function SputnikError(message)
{
    this.message = message;
}

SputnikError.prototype.toString = function ()
{
    return 'SputnikError: ' + this.message;
};

var sputnikException;

function testPrint(msg)
{
    var span = document.createElement("span");
    document.getElementById("console").appendChild(span); // insert it first so XHTML knows the namespace 
    span.innerHTML = msg + '<br />';
}

function escapeHTML(text)
{
    return text.toString().replace(/&/g, "&amp;").replace(/</g, "&lt;");
}

function printTestPassed(msg)
{
    testPrint('<span><span class="pass">PASS</span> ' + escapeHTML(msg) + '</span>');
}

function printTestFailed(msg)
{
    testPrint('<span><span class="fail">FAIL</span> ' + escapeHTML(msg) + '</span>');
}

function testFailed(msg)
{
    throw new SputnikError(msg);
}

var successfullyParsed = false;
</script>

</head>
<body>
<p>S12.7_A8_T1</p>
<div id='console'></div>
<script>
try {

/**
* @name: S12.7_A8_T1;
* @section: 12.7;
* @assertion: Appearing of "continue" within a "try/catch" Block yields SyntaxError;
* @description: Checking if execution of "continue Identifier" within catch Block fails;
* @negative;
*/

var x=0,y=0;

try{
	LABEL1 : do {
		x++;
		throw "gonna leave it";
		y++;
	} while(0);
	testFailed('#1: throw "gonna leave it" lead to throwing exception');
} catch(e){
	continue LABEL2;
	LABEL2 : do {
		x++;
		y++;
	} while(0);
};

} catch (ex) {
    sputnikException = ex;
}

var successfullyParsed = true;
</script>

<script>
if (!successfullyParsed)
    printTestPassed('Expected parsing failure');
else if (sputnikException)
    printTestPassed(sputnikException);
else
    printTestFailed("No error detected");
testPrint('<br /><span class="pass">TEST COMPLETE</span>');
</script>
</body>
</html>
